requirements.txt
playwright
sentence-transformers
torch
faiss-cpu
groq
streamlit
numpy